requirements.txt
groq==0.13.1
Flask==3.1.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
click==8.1.7